+++ matter_server/common/model/message.py
@@ -1,5 +1,9 @@
+"""Models for the Websocket messages."""
+from __future__ import annotations
+
 from dataclasses import dataclass
 from typing import Any
+from enum import Enum, IntEnum
 
 
 @dataclass
@@ -10,7 +14,6 @@ class Message:
 @dataclass
 class ResultMessage:
     messageId: str
-    pass
 
 
 @dataclass
@@ -18,9 +21,18 @@ class SuccessResultMessage(ResultMessage):
     result: Any
 
 
+class ErrorCode(IntEnum):
+    """Enum with possible error codes."""
+
+    INVALID_COMMAND = 1
+    NOT_FOUND = 2
+    UNKNOWN_ERROR = 99
+
+
 @dataclass
 class ErrorResultMessage(ResultMessage):
-    errorCode: Any
+    errorCode: ErrorCode
+    details: str | None = None
 
 
 @dataclass
@@ -40,3 +52,19 @@ class CommandMessage(Message):
 class ServerInformation:
     fabricId: int
     compressedFabricId: int
+
+
+class EventType(Enum):
+    """Enum with possible events sent from server to client."""
+
+    DEVICE_ADDED = "device_added"
+    DEVICE_UPDATED = "device_updated"
+    DEVICE_DELETED = "device_deleted"
+
+
+@dataclass
+class ServerEvent:
+    """Event sent from server to the client."""
+
+    type: EventType
+    data: Any
