+++ matter_server/server/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of a Websocket-based Matter proxy (using CHIP SDK)."""
